tests/common/cloudrun/requirements.txt
requests==2.26.0
google-cloud==0.34.0
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
google-resumable-media==2.7.0
workflow
python-dotenv==0.20.0
protobuf==3.20.*